backend/requirements.txt
# Core Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.0

# Database
sqlalchemy==2.0.36
alembic==1.14.0
psycopg[binary]==3.2.3

# Redis
redis==5.2.1
hiredis==3.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
bcrypt==4.2.1

# Email & SMS
sendgrid==6.11.0
requests==2.32.3

# Payments
razorpay==1.4.2

# Utilities
python-dotenv==1.0.1
pytz==2024.2
phonenumbers==8.13.51

# Development
pytest==8.3.4
pytest-asyncio==0.24.0
black==24.10.0
flake8==7.1.1
httpx==0.28.1
